+++ src/rpc/src/IdGenerator/RequestIdGenerator.php
@@ -10,13 +10,15 @@ declare(strict_types=1);
  * @license  https://github.com/hyperf-cloud/hyperf/blob/master/LICENSE
  */
 
-namespace Hyperf\Rpc;
+namespace Hyperf\Rpc\IdGenerator;
 
-class RequestIdGenerator
+use Hyperf\Contract\IdGeneratorInterface;
+
+class RequestIdGenerator implements IdGeneratorInterface
 {
-    public function generate(): int
+    public function generate(): string
     {
         $us = strstr(microtime(), ' ', true);
-        return intval(strval($us * 1000 * 1000) . rand(100, 999));
+        return strval($us * 1000 * 1000) . rand(100, 999);
     }
 }
